+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - `livechatconnector/v3/getchattoken` endpoint
 - `livechatconnector/v3/auth/getchattoken` endpoint
+- Stop retry when the error is related to out of office hours.
 
 ## [0.3.3] - 2023-01-09
 ### Fix